--- src/epics/FirebaseChatEpic.js
@@ -1,1 +1,0 @@
-
